topaz: Patch Goodix fingerprint HAL to ignore conflicting events
This will fix the long standing issue of the screenshot button
combination not working properly in AOSP for Goodix sensors.

The patch will replace the first instruction of the
'gf_hal_send_key_event' function with a ret so that the touch
events which conflict with AOSP do not get reported to userspace.

https://github.com/LineageOS/android_device_xiaomi_rosemary/commit/59e0aa7dec751def8c679edd6de2abb67495002f

Signed-off-by: Rama Bondan Prakoso <ramanarubp@gmail.com>